% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdChicken.R
\docType{data}
\name{BirdChicken}
\alias{BirdChicken}
\title{Bir/Chicken shape classification}
\format{\describe{
The variables are as follows:
\itemize{
\item{\code{df}:}{ \code{data.frame} with the following variables:}
\itemize{
\item{\code{class}:}{ Corresponding class level of ``BirdChicken'' curves with 2 classes:level 1, n=20 and level 2, n=20.}
\item{\code{sample}:}{Factor variable. In TSC, the first 20 values (\code{sample=train}) are used for training sample and the rest of 20 (\code{sample=test}) for testing.}
}
\item{\code{x}:}{ \code{fdata} class object with with n=40 curves (per row) observed in 512 discretization points (per column).}
}

}}
\source{
MPEG-7 CE Shape-1 Part B is a database of binary images developed for testing MPEG-7 shape descriptors, and is available free online: \url{http://www.dabi.temple.edu/~shape/MPEG7/dataset.html} and
\url{http://timeseriesclassification.com/description.php?Dataset=BirdChicken}
}
\usage{
data(BirdChicken)
}
\description{
It is used for testing contour/image and skeleton-based descriptors. Classes of images vary broadly, and include classes that are similar in shape to one another.
}
\details{
There are 20 instances of each class, and 60 classes in total. We have extracted the outlines of these images and mapped them into 1-D series of distances to the centre. Bird/Chicken is the problem of distinguishing between an outline of a bird and a chicken
}
\keyword{datasets}
