% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coffee.R
\docType{data}
\name{Coffee}
\alias{Coffee}
\title{Coffee data set}
\format{\describe{
The variables are as follows:
\itemize{
\item{\code{df}:}{ \code{data.frame} with the following variables:}
\itemize{
\item{\code{class}:}{ Corresponding class level of ``Coffee'' curves with 2 classes.}
\item{\code{sample}:}{Factor variable. In TSC database, the first 28 values (\code{sample=train}) are used for training sample and the rest of 28 (\code{sample=test}) for testing.}
}
\item{\code{x}:}{ \code{fdata} class object with with n=56 curves (per row) in 286 discretization points (per column).}
}

}}
\source{
\url{https://csr.quadram.ac.uk/example-datasets-for-download/} and \url{http://timeseriesclassification.com/description.php?Dataset=Coffee}
}
\usage{
data(Coffee)
}
\description{
Food spectrographs are used in chemometrics to classify food types, a task that has obvious applications in food safety and quality assurance. The coffee data set is a two class problem to distinguish between Robusta and Aribica coffee beans.
}
\details{
Further information can be found in the original paper Briandet et al. Discrimination of Arabica and Robusta in Instant Coffee by Fourier Transform Infrared Spectroscopy and Chemometrics J. Agricultural and Food Chemistry, 44 (1), 1996. The data was first used in the time series classification literature in Bagnall et al. Transformation Based Ensembles for Time Series Classification, SDM 2012.
}
\keyword{datasets}
