% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsectWingbeat.R
\docType{data}
\name{InsectWingbeatSoundTSC}
\alias{InsectWingbeatSoundTSC}
\title{Flying Insect Classification with Inexpensive Sensors}
\format{\describe{
The variables are as follows:
\itemize{
\item{\code{df}:}{ \code{data.frame} with the following variables:}
\itemize{
\item{\code{class}:}{ Corresponding class level of ``InsectWingbeatSound'' curves with 11 classes.}
\item{\code{sample}:}{Factor variable. In TSC database, the first 220 values (\code{sample=train}) are used for training sample and the rest of 1980 (\code{sample=test}) for testing.}
}
\item{\code{x}:}{ \code{fdata} class object with with n=2200 curves (per row) in 256 discretization points (per column).}
}

}}
\source{
\url{https://sites.google.com/site/insectclassification/} and \url{http://timeseriesclassification.com/description.php?Dataset=InsectWingbeatSound}
}
\usage{
data(InsectWingbeatSoundTSC)
}
\description{
The original data is a reconstruction of the sound of insects passing through a sensor. The data in the archive is the power spectrum of the sound. The 11 classes are male and female mosquitos (Ae. aegypti, Cx. tarsalis, Cx. quinquefasciants, Cx. stigmatosoma), two types of flies (Musca domestica and Drosophila simulans) and other insects.
}
\details{
The InsectWingbeatSound data was generated by the UCR computational entomology group and used in the paper Flying Insect Classification with Inexpensive Sensors. Journal of Insect Behaviour 2014.
}
\references{
The conditions data was gathered in can be found at \url{http://alumni.cs.ucr.edu/~ychen053/InsectBehaviour_059.pdf}
}
\keyword{datasets}
