% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmallKitchenAppliances.R
\docType{data}
\name{SmallKitchenAppliances}
\alias{SmallKitchenAppliances}
\title{Small Kitchen Appliances}
\format{\describe{
The variables are as follows:
\itemize{
\item{\code{df}:}{ \code{data.frame} with the following variables:}
\itemize{
\item{\code{class}:}{ Corresponding class level of ``SmallKitchenAppliances'' curves with 3 classes (250 observations per class).}
\item{\code{sample}:}{Factor variable. In TSC database, the first 375 values (\code{sample=train}) are used for training sample and the rest of 375 (\code{sample=test}) for testing.}
}
\item{\code{x}:}{ \code{fdata} class object with with n=750 curves (per row) in 720 discretization points (per column).}
}

}}
\source{
\url{https://www.uea.ac.uk/computing/machine-learning/elastic-ensembles} and \url{http://timeseriesclassification.com/description.php?Dataset=SmallKitchenAppliances}
}
\usage{
data(SmallKitchenAppliances)
}
\description{
The data contains readings from 251 households, sampled in two-minute intervals over a month. Each series is length 720 (24 hours of readings taken every 2 minutes). Classes areKettle, Microwave and Toaster
}
\details{
These problems were taken from data recorded as part of government sponsored study called Powering the Nation. The intention was to collect behavioural data about how consumers use electricity within the home to help reduce the UK's carbon footprint.
}
\keyword{datasets}
